test(core): extract the HookRegistry unit tests from plugin-registry.test.ts
plugin-registry.test.ts sat at 1,212 lines — right at the SPEC ceiling —
and its "Section A" tested a different module entirely (packages/core's
HookRegistry class, needing none of the registry's mock scaffold). That
section moves verbatim to tests/core/hook-registry.test.ts (104 lines);
plugin-registry.test.ts (1,131) now holds only the PluginRegistryImpl
suite, whose inner describes all share one activation harness and stay
together as a cohesive unit.

Pure relocation: 52 pass / 99 expect() calls across the pair, unchanged.
